test: completes use cases tests

+++ internal/human/models/stats.go
@@ -4,5 +4,5 @@ package models
 type Stats struct {
 	CountMutantDNA int16   `json:"count_mutant_dna"`
 	CountHumanDNA  int16   `json:"count_human_dna"`
-	Ratio          float32 `json:"ratio"`
+	Ratio          float64 `json:"ratio"`
 }
